fix (DS1882): compilation error on some arch due to unspecified type

+++ src/audio/DS1882.h
@@ -31,13 +31,13 @@
   #define DS1882_MAXIMUM_VOL_LEVEL    ((uint8_t)63U)  // cast for Mbed compatibility
 
   // use volatile or non-volatile memory storage
-  #define USE_VOLATILE_MEMORY_STORAGE ( 0U)
+  #define USE_VOLATILE_MEMORY_STORAGE ((uint8_t)0U)
 
   // enable ZCD to reduce artifacts when changing volume levels
-  #define ENABLE_ZERO_CROSSING_DETECT ( 0U)
+  #define ENABLE_ZERO_CROSSING_DETECT ((uint8_t)0U)
 
   // potentiometer configuration option, must be set to 1
-  #define POTENTIOMETER_CONFIG_OPTION ( 1U)
+  #define POTENTIOMETER_CONFIG_OPTION ((uint8_t)1U)
 
   // Combine configuration register options into one single value
   #define DS1882_CONFIGURATION (0x80 + (USE_VOLATILE_MEMORY_STORAGE << 2) + (ENABLE_ZERO_CROSSING_DETECT << 1) + (POTENTIOMETER_CONFIG_OPTION - 1))
